fix gss insert and search and update data structures
fixing gss algorithm to avoid always adding into leftover and replace
some vectors to static array

+++ gss/LinkedList.cpp
@@ -2,7 +2,7 @@
 
 LinkedList::LinkedList(
     ull addr, 
-    int weigth):
+    ull weigth):
     addr(addr), 
     weigth(weigth), 
     next(nullptr) {}
