+++ gss/LinkedList.hpp
@@ -4,7 +4,7 @@ class LinkedList {
 public:
     LinkedList* next;
     ull addr;
-    int weigth;
+    ull weigth;
 
-    LinkedList(ull addr, int weigth);
+    LinkedList(ull addr, ull weigth);
 };
